[CLUB-32] create service pages
